// src/app/connectors/api/model/registerRequestWeb.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterRequestWeb { 
    email?: string;
    password?: string;
    name?: string;
    lastName?: string;
    birthDate?: string;
}